+++ build.assets/tooling/cmd/check/main.go
@@ -30,9 +30,8 @@ import (
 
 	"golang.org/x/mod/semver"
 
+	"github.com/gravitational/teleport/build.assets/tooling/lib/github"
 	"github.com/gravitational/trace"
-
-	go_github "github.com/google/go-github/v41/github"
 )
 
 func main() {
@@ -46,7 +45,7 @@ func main() {
 
 	switch check {
 	case "latest":
-		err = checkLatest(ctx, tag, newGitHub())
+		err = checkLatest(ctx, tag, github.NewGitHub())
 	case "prerelease":
 		err = checkPrerelease(tag)
 	default:
@@ -78,7 +77,7 @@ func parseFlags() (string, string, error) {
 	return *tag, *check, nil
 }
 
-func checkLatest(ctx context.Context, tag string, gh github) error {
+func checkLatest(ctx context.Context, tag string, gh github.GitHub) error {
 	releases, err := gh.ListReleases(ctx, "gravitational", "teleport")
 	if err != nil {
 		return trace.Wrap(err)
@@ -105,47 +104,4 @@ func checkPrerelease(tag string) error {
 		return trace.BadParameter("version contains build metadata: %v", tag)
 	}
 	return nil
-}
-
-type github interface {
-	ListReleases(ctx context.Context, organization string, repository string) ([]string, error)
-}
-
-type ghClient struct {
-	client *go_github.Client
-}
-
-func newGitHub() *ghClient {
-	return &ghClient{
-		client: go_github.NewClient(nil),
-	}
-}
-
-func (c *ghClient) ListReleases(ctx context.Context, organization string, repository string) ([]string, error) {
-	var releases []string
-
-	opt := &go_github.ListOptions{
-		Page:    0,
-		PerPage: 100,
-	}
-	for n := 0; n < 100; n++ {
-		page, resp, err := c.client.Repositories.ListReleases(ctx,
-			organization,
-			repository,
-			opt)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-
-		for _, p := range page {
-			releases = append(releases, p.GetTagName())
-		}
-
-		if resp.NextPage == 0 {
-			break
-		}
-		opt.Page = resp.NextPage
-	}
-
-	return releases, nil
 }
