+++ build.assets/tooling/go.mod
@@ -1,17 +1,20 @@
-module github.com/gravitational/teleport/build.assets/version-check
+module github.com/gravitational/teleport/build.assets/tooling
 
 go 1.17
 
-require github.com/google/go-github/v41 v41.0.0
+require (
+	github.com/google/go-github/v41 v41.0.0
+	github.com/gravitational/trace v1.1.15
+	golang.org/x/mod v0.5.1
+)
 
 require (
 	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/gravitational/trace v1.1.15 // indirect
 	github.com/jonboulle/clockwork v0.2.2 // indirect
 	github.com/sirupsen/logrus v1.8.1 // indirect
 	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
-	golang.org/x/mod v0.5.1 // indirect
 	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
 	golang.org/x/term v0.0.0-20201126162022-7de9c90e9dd1 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
